requirements.txt
discord.py[voice]
emoji
ffmpeg-python
ffmpy
gTTS